refactor : review exception 정리 및 서비스 메소드 추출
