+++ src/main/java/com/ybe/tr1ll1on/domain/review/exception/ReviewException.java
@@ -3,6 +3,8 @@ package com.ybe.tr1ll1on.domain.review.exception;
 import com.ybe.tr1ll1on.global.exception.ExceptionCode;
 import com.ybe.tr1ll1on.global.exception.TrillionException;
 
-public class ReviewAlreadyWrittenException extends TrillionException {
-    public ReviewAlreadyWrittenException (ExceptionCode errorCode) { super(errorCode); }
+public class ReviewException extends TrillionException {
+    public ReviewException(ExceptionCode exceptionCode) {
+        super(exceptionCode);
+    }
 }
